refactor funcionality of edit and create page
